Let GitHub Actions assist for release builds
By leveraging the proven action for snapshot builds, the manual and
potentially error-prone releasing building is simplified.

Note that releases are not deployed automatically to SourceForge. They
are temporarily stored as build artifact on GitHub. This gives the
release manager the final decision to manually upload the new release
to SourceForge after a final check.